% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_500.R
\name{get_500}
\alias{get_500}
\title{get_500}
\usage{
get_500(time, subreddit)
}
\arguments{
\item{time}{Epoch time for the scraping to begin}

\item{subreddit}{A valid subreddit name}
}
\value{
The titles of the 500 post previous to the input time of the subreddit.
}
\description{
This function takes an epoch time and a subreddit name and returns the titles of the 500 posts to that subreddit before the time. It also returns the time (epoch) of the last post scraped
}
\examples{
subreddit <- "rabbits"
times <- 1558699200   #24/05/2019 12:00:00 pm in epoch time

titles <- get_500(times,subreddit)

}
