% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_word_freq.R
\name{get_word_freq}
\alias{get_word_freq}
\title{Get Title word frequency}
\usage{
get_word_freq(num_posts, subreddit, num_words = 100,
  colour_pallete = "Dark2", num_colours = 8)
}
\arguments{
\item{num_posts}{Number of reddit posts to be scraped}

\item{subreddit}{The subreddit to be scraped}

\item{num_words}{The number of words to be plotted in the wordcloud}

\item{colour_pallete}{The colour_pallete to be used in the wordcloud}

\item{num_colours}{The number of different colours to be used in the wordcloud}
}
\value{
A data frame with the 100 most common words and their frequencies and a plot of the wordcloud
}
\description{
This function takes a subreddit and the number of posts we wish to scrape, and returns a data frame with the most common words in the title of the posts and their respective frequencies and a plot of the wordcloud.
}
\examples{
num_posts = 10000;
subreddit = "rabbits"

#Default plot
words_freq <- get_word_freq(num_posts, subreddit)

#Custom plot
get_word_freq <- get_word_freq(num_posts, subreddit,
                                num_words = 50,
                                colour_pallete = "Set1",
                                num_colours = 5)
}
