% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CircketData.R
\docType{data}
\name{CricketData}
\alias{CricketData}
\title{Statistics for cricket World cup from 1971 to 2017}
\format{A data frame with 3932 rows and 7 variables:
\describe{
 \item{Team 1}{The home team}
 \item{Team 2}{The away team}
 \item{Winner}{The away team}
 \item{Margin}{The winning margin (in cricket terminology)}
 \item{Ground}{The ground the match was played at}
 \item{Match Date}{The date the game was plated, month (abv) day, year.}
}}
\source{
\url{JoshuaBean}
}
\usage{
CricketData
}
\description{
A dataset containing Statistic for each game in crticket World Cup from 1971 to 2017.
}
\keyword{datasets}
